more

+++ tcc_boot_min/tcc.h
@@ -143,7 +143,7 @@ typedef union CValue {
 /* value on stack */
 typedef struct SValue {
     CType type;      /* type */
-    unsigned short r;      /* register + flags */
+    int r;      /* register + flags */
     CValue c;              /* constant, if VT_CONST */
     struct Sym *sym;       /* symbol, if (VT_SYM | VT_CONST), or if
     			      result of unary() for an identifier. */
